==> resource_tracking/prep/2g_gf_visualize_data.rmd <==
---
title: "Resource Tracking Descriptive Verification"
author: "Emily Linebarger"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: pdf_document
---

```{r setup, warning=FALSE, echo=FALSE, message=FALSE}
library(knitr)
library(data.table)
library(ggplot2)
library(lubridate)

knitr::opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, fig.width=11, fig.height=8)

budgets = readRDS(paste0(box, "tableau_data/final_budgets.rds"))
expenditures = readRDS(paste0(box, "tableau_data/final_expenditures.rds"))

absorption = readRDS(paste0(box, "tableau_data/absorption.rds"))
sen_absorption = absorption[loc_name=="sen"]
cod_absorption = absorption[loc_name=="cod"]
gtm_absorption = absorption[loc_name=="gtm"]
uga_absorption = absorption[loc_name=="uga"]

```

# BUDGETS AND EXPENDITURES 

## Plot data
Visualize the budget data by grant and year. 
Focus only on current grant first. 
```{r plot1}
sum_current_grants = budgets[, .(budget=sum(budget, na.rm = TRUE)), by=.(grant, year, loc_name)]
grants = unique(current_uga_grants)
for (grant_number in grants){
current_uga = ggplot(sum_current_grants[grant==grant_number & year>=2015], aes(x=year, y=budget)) + 
  geom_bar(stat="identity", fill="forestgreen")+
  theme_bw() + labs(title=paste0("Sum of budget per year for ", grant_number), x="Year", y="Budget", 
                    caption="*Years limited to 2015-present")
print(current_uga)
print(sum_current_grants[grant==grant_number])
}

grants = unique(current_cod_grants)
for (grant_number in grants){
current_cod = ggplot(sum_current_grants[grant==grant_number & year>=2015], aes(x=year, y=budget)) + 
  geom_bar(stat="identity", fill="blue")+
  theme_bw() + labs(title=paste0("Sum of budget per year for ", grant_number), x="Year", y="Budget", 
                    caption="*Years limited to 2015-present")
print(current_cod)
print(sum_current_grants[grant==grant_number])
}

grants = unique(current_gtm_grants)
for (grant_number in grants){
current_gtm = ggplot(sum_current_grants[grant==grant_number & year>=2015], aes(x=year, y=budget)) + 
  geom_bar(stat="identity", fill="purple")+
  theme_bw() + labs(title=paste0("Sum of budget per year for ", grant_number), x="Year", y="Budget", 
                    caption="*Years limited to 2015-present")
print(current_gtm)
print(sum_current_grants[grant==grant_number])
}


```

## Visualize by disease and country 
```{r plot2}
sum_by_country_disease = budgets[, .(budget=sum(budget, na.rm=TRUE)), by=c('country', 'year', 'disease')]
countries = unique(budgets$country)
for (loc in countries){
hiv = ggplot(sum_by_country_disease[country==loc & disease=='hiv'], aes(x=year, y=budget)) + 
  geom_bar(stat="identity", fill="red")+
  theme_bw() + labs(title=paste0("Sum of budget per year for ", loc, " for HIV"), x="Year", y="Budget")
print(hiv)

tb = ggplot(sum_by_country_disease[country==loc & disease=='tb'], aes(x=year, y=budget)) + 
  geom_bar(stat="identity", fill="blue")+
  theme_bw() + labs(title=paste0("Sum of budget per year for ", loc, " for TB"), x="Year", y="Budget")
print(tb)

malaria = ggplot(sum_by_country_disease[country==loc & disease=='malaria'], aes(x=year, y=budget)) + 
  geom_bar(stat="identity", fill="yellow")+
  theme_bw() + labs(title=paste0("Sum of budget per year for ", loc, " for Malaria"), x="Year", y="Budget")
print(malaria)

rssh = ggplot(sum_by_country_disease[country==loc & disease=='rssh'], aes(x=year, y=budget)) + 
  geom_bar(stat="identity", fill="purple")+
  theme_bw() + labs(title=paste0("Sum of budget per year for ", loc, " for RSSH"), x="Year", y="Budget")
print(rssh)
}

```

## Review variables 

Are the formats of the variables what you expect? 
```{r plot3}
# str(budgets)
# str(expenditures)
```

Check specific variable values 
```{r plot4}

disease_flag_b = (unique(budgets$disease)%in%c('hiv', 'malaria', 'rssh', 'tb'))
disease_flag_p = (unique(expenditures$disease)%in%c('hiv', 'malaria', 'rssh', 'tb'))

sdate_flag_b = is.Date(budgets$start_date)
sdate_flag_p = is.Date(expenditures$start_date)

edate_flag_b = is.Date(budgets$end_date)
edate_flag_p = is.Date(expenditures$end_date)




```

Are you returning only expected values for 'disease' column in budgets?: `r disease_flag_b`
Are you returning only expected values for 'disease' column in PUDRs?: `r disease_flag_p`

Is your start date variable a date in budgets? `r sdate_flag_b`
What about your end date variable? `r edate_flag_b`
What's the range of start date? `r range(budgets$start_date)`

Is your start date variable a date in PUDRs? `r sdate_flag_p`
What about your end date variable? `r edate_flag_p`
What's the range of start date? `r range(expenditures$start_date)`

What's the range of years in your budgets dataset? $`r range(budgets$year)`$
What's the range of years in your expenditures dataset? $`r range(expenditures$year)`$

What values do you have for language, loc_name, and country?
For budgets: 
```{r plot5}
unique(budgets$language)
unique(budgets$loc_name)
unique(budgets$country)
```
For expenditures: 
```{r plot6}
unique(expenditures$language)
unique(expenditures$loc_name)
unique(expenditures$country)
```

Is the 'current grant' flag working correctly? Make sure all files are sorted by this boolean correctly. 
```{r plot7}
unique(budgets[current_grant==TRUE, .(current_grant, grant, grant_period)][order(grant, grant_period)])
unique(budgets[current_grant==FALSE, .(current_grant, grant, grant_period)][order(grant, grant_period)])
```

Is the 'grant_disease' flag working correctly? Make sure that the disease makes sense for the name of the grant. 
First showing budget and then expenditure. 
```{r plot8}
unique(budgets[, .(grant, grant_disease)])
unique(expenditures[, .(grant, grant_disease)])
```

# ABSORPTION GRAPHS

```{r cod_absorption} 
absorption_by_country = cod_absorption[, .(budget=sum(budget, na.rm=T), expenditure=sum(expenditure, na.rm=T)), by=c('grant', 'grant_period', 'semester')]
absorption_by_country[, absorption:=(expenditure/budget)*100]

#Where you don't have data for a semester, show it as NA. 
all_semesters = expand.grid(semester=unique(absorption_by_country$semester), grant=unique(absorption_by_country$grant))
absorption_by_country = merge(absorption_by_country, all_semesters, all.y=T, by=c('grant', 'semester')) 

by_grant = ggplot(absorption_by_country, aes(x=grant, y=absorption, group=semester, fill=semester)) + 
  geom_bar(stat="identity", position="dodge")+
  theme_bw()+
  labs(title="DRC Absorption by grant and semester", y="Absorption (%)", x="Grant")

by_grant

```

```{r gtm_absorption} 
absorption_by_country = gtm_absorption[, .(budget=sum(budget, na.rm=T), expenditure=sum(expenditure, na.rm=T)), by=c('grant', 'grant_period', 'semester')]
absorption_by_country[, absorption:=(expenditure/budget)*100]

#Where you don't have data for a semester, show it as NA. 
all_semesters = expand.grid(semester=unique(absorption_by_country$semester), grant=unique(absorption_by_country$grant))
absorption_by_country = merge(absorption_by_country, all_semesters, all.y=T, by=c('grant', 'semester')) 

by_grant = ggplot(absorption_by_country, aes(x=grant, y=absorption, group=semester, fill=semester)) + 
  geom_bar(stat="identity", position="dodge")+
  theme_bw()+
  labs(title="Guatemala Absorption by grant and semester", y="Absorption (%)", x="Grant") 

by_grant

```

```{r sen_absorption} 
absorption_by_country = sen_absorption[, .(budget=sum(budget, na.rm=T), expenditure=sum(expenditure, na.rm=T)), by=c('grant', 'grant_period', 'semester')]
absorption_by_country[, absorption:=(expenditure/budget)*100]

#Where you don't have data for a semester, show it as NA. 
all_semesters = expand.grid(semester=unique(absorption_by_country$semester), grant=unique(absorption_by_country$grant))
absorption_by_country = merge(absorption_by_country, all_semesters, all.y=T, by=c('grant', 'semester')) 

by_grant = ggplot(absorption_by_country, aes(x=grant, y=absorption, group=semester, fill=semester)) + 
  geom_bar(stat="identity", position="dodge")+
  theme_bw()+
  labs(title="Senegal Absorption by grant and semester", y="Absorption (%)", x="Grant")

by_grant

```

```{r uga_absorption} 
absorption_by_country = uga_absorption[, .(budget=sum(budget, na.rm=T), expenditure=sum(expenditure, na.rm=T)), by=c('grant', 'grant_period', 'semester')]
absorption_by_country[, absorption:=(expenditure/budget)*100]

#Where you don't have data for a semester, show it as NA. 
all_semesters = expand.grid(semester=unique(absorption_by_country$semester), grant=unique(absorption_by_country$grant))
absorption_by_country = merge(absorption_by_country, all_semesters, all.y=T, by=c('grant', 'semester')) 

by_grant = ggplot(absorption_by_country, aes(x=grant, y=absorption, group=semester, fill=semester)) + 
  geom_bar(stat="identity", position="dodge")+
  theme_bw()+
  labs(title="Uganda Absorption by grant and semester", y="Absorption (%)", x="Grant") 

by_grant

```
